+++ src/loggers/index.ts
@@ -1,2 +1,3 @@
 export * from './BrowserLogger';
+export * from './PlainLogger';
 export * from './ChalkLogger';